moved back to ints from floats, not even sure why fixed lines being created in red/cyan lacking filter moved overlap generation to only occurring when dots are hidden TODO: figure out why lines go 1 px thin when vertical or horizontal TODO: figure out why action during red/cyan is still slower even when no overlaps are present

+++ StereoSketcher/selectangle.js
@@ -4,10 +4,10 @@ function changeSelectangle(event) {
 	event.stopPropagation();
 	if(selectangle!=null)
 	{
-		var x=parseFloat(selectangle.getAttribute("x"));
-		var y=parseFloat(selectangle.getAttribute("y"));
-		var width=parseFloat(selectangle.getAttribute("width"));
-		var height=parseFloat(selectangle.getAttribute("height"));
+		var x=parseInt(selectangle.getAttribute("x"));
+		var y=parseInt(selectangle.getAttribute("y"));
+		var width=parseInt(selectangle.getAttribute("width"));
+		var height=parseInt(selectangle.getAttribute("height"));
 		var ex=event.clientX;
 		var ey=event.clientY;
 		if(selectangle.originalX<ex)
